win32 support

+++ ide.c
@@ -394,7 +394,7 @@ static void padstr(char *str, const char *src, int len)
             v = *src++;
         else
             v = ' ';
-        *(char *)((long)str ^ 1) = v;
+        *(char *)((uintptr_t)str ^ 1) = v;
         str++;
     }
 }
